feat(DeploymentEndpoint): moved utility methods to DeploymentUtilites class
